requirements.txt
fastapi>=0.115
uvicorn>=0.34
python-dotenv>=1.0
psycopg2-binary>=2.9
jinja2>=3.1
python-multipart>=0.0.9
